# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Agent Planner API",
    "version": "2.0.0",
    "description": "Collaborative planning API for humans and AI agents"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization"
      }
    },
    "schemas": {
...
      "DecisionOption": {
        "type": "object",
        "properties": {
          "option": {
            "type": "string",
            "description": "The option being proposed"
          },
          "pros": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Advantages of this option"
          },
          "cons": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Disadvantages of this option"
          },
          "recommendation": {
            "type": "boolean",
            "description": "Whether this is the recommended option"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/plans/{id}/decisions": {
      "get": {
        "summary": "List decision requests for a plan",
        "tags": [
          "Decisions"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Plan UUID"
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "pending",
                "decided",
                "expired",
                "cancelled"
              ]
            },
            "description": "Filter by status"
          },
          {
            "in": "query",
            "name": "urgency",
            "schema": {
              "type": "string",
              "enum": [
                "blocking",
                "can_continue",
                "informational"
              ]
            },
            "description": "Filter by urgency"
          },
          {
            "in": "query",
            "name": "node_id",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Filter by node"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 50
            }
          },
          {
            "in": "query",
            "name": "offset",
            "schema": {
              "type": "integer",
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of decision requests",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DecisionRequest"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new decision request",
        "tags": [
          "Decisions"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "title",
                  "context"
                ],
                "properties": {
                  "node_id": {
                    "type": "string",
                    "format": "uuid",
                    "description": "Optional node this decision relates to"
                  },
                  "title": {
                    "type": "string",
                    "maxLength": 200,
                    "description": "Brief title for the decision"
                  },
                  "context": {
                    "type": "string",
                    "maxLength": 5000,
                    "description": "Full context explaining what needs to be decided"
                  },
                  "options": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/DecisionOption"
                    },
                    "description": "Structured options with pros/cons"
                  },
                  "urgency": {
                    "type": "string",
                    "enum": [
                      "blocking",
                      "can_continue",
                      "informational"
                    ],
                    "default": "can_continue"
                  },
                  "expires_at": {
                    "type": "string",
                    "format": "date-time",
                    "description": "Optional expiration timestamp"
                  },
                  "requested_by_agent_name": {
                    "type": "string",
                    "maxLength": 100,
                    "description": "Name of the agent requesting"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Decision request created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionRequest"
                }
              }
            }
          }
        }
      }
    },
...
    "/plans/{planId}/nodes/{nodeId}/dependencies": {
      "get": {
        "summary": "List dependencies for a specific node",
        "tags": [
          "Dependencies"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "planId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "in": "path",
            "name": "nodeId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "in": "query",
            "name": "direction",
            "schema": {
              "type": "string",
              "enum": [
                "upstream",
                "downstream",
                "both"
              ],
              "default": "both"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Node dependencies"
          }
        }
      }
    },
...
    "/plans/{id}/nodes": {
      "get": {
        "summary": "Get all nodes for a plan (tree structure)",
        "description": "Returns hierarchical tree of nodes. By default returns minimal fields (id, parent_id, node_type, title, status, order_index) for efficient navigation. Use include_details=true for full node data, or use GET /plans/{id}/nodes/{nodeId}/context for detailed information about specific nodes.",
        "tags": [
          "Nodes"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The plan ID"
          },
          {
            "in": "query",
            "name": "include_details",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false
            },
            "description": "Include full node details (description, context, agent_instructions, metadata, timestamps). Default is false for minimal response size."
          },
          {
            "in": "query",
            "name": "coherence_status",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "coherent",
                "stale_beliefs",
                "contradiction_detected",
                "unchecked"
              ]
            },
            "description": "Filter nodes by BDI coherence status. Supports comma-separated values."
          }
        ],
        "responses": {
          "200": {
            "description": "Hierarchical tree of plan nodes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Node"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Access denied"
          },
          "404": {
            "description": "Plan not found"
          }
        }
      },
      "post": {
        "summary": "Create a new node in a plan",
        "tags": [
          "Nodes"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The plan ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "node_type",
                  "title"
                ],
                "properties": {
                  "parent_id": {
                    "type": "string",
                    "description": "Parent node ID (if not provided, will be assigned to the root node)"
                  },
                  "node_type": {
                    "type": "string",
                    "enum": [
                      "phase",
                      "task",
                      "milestone"
                    ]
                  },
                  "title": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "not_started",
                      "in_progress",
                      "completed",
                      "blocked"
                    ]
                  },
                  "order_index": {
                    "type": "integer"
                  },
                  "due_date": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "context": {
                    "type": "string"
                  },
                  "agent_instructions": {
                    "type": "string"
                  },
                  "metadata": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Node created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Node"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Access denied"
          },
          "404": {
            "description": "Plan or parent node not found"
          }
        }
      }
    },
    "/plans/{id}/nodes/{nodeId}": {
      "get": {
        "summary": "Get a specific node",
        "tags": [
          "Nodes"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The plan ID"
          },
          {
            "in": "path",
            "name": "nodeId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The node ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Node details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Node"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Access denied"
          },
          "404": {
            "description": "Node not found"
          }
        }
      },
      "put": {
        "summary": "Update a node",
        "tags": [
          "Nodes"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The plan ID"
          },
          {
            "in": "path",
            "name": "nodeId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The node ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "node_type": {
                    "type": "string",
                    "enum": [
                      "phase",
                      "task",
                      "milestone"
                    ]
                  },
                  "title": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "not_started",
                      "in_progress",
                      "completed",
                      "blocked"
                    ]
                  },
                  "order_index": {
                    "type": "integer"
                  },
                  "due_date": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "context": {
                    "type": "string"
                  },
                  "agent_instructions": {
                    "type": "string"
                  },
                  "metadata": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Node updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Node"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Access denied"
          },
          "404": {
            "description": "Node not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a node",
        "tags": [
          "Nodes"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The plan ID"
          },
          {
            "in": "path",
            "name": "nodeId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The node ID"
          }
        ],
        "responses": {
          "204": {
            "description": "Node deleted successfully"
          },
          "400": {
            "description": "Cannot delete root node"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Access denied"
          },
          "404": {
            "description": "Node not found"
          }
        }
      }
    },
...
    "/plans/{id}/nodes/{nodeId}/move": {
      "post": {
        "summary": "Move a node to a different parent or position",
        "tags": [
          "Nodes"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The plan ID"
          },
          {
            "in": "path",
            "name": "nodeId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The node ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "parent_id": {
                    "type": "string",
                    "description": "New parent node ID"
                  },
                  "order_index": {
                    "type": "integer",
                    "description": "New position among siblings"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Node moved successfully"
          },
          "400": {
            "description": "Invalid input or cannot move root node"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Access denied"
          },
          "404": {
            "description": "Node or parent not found"
          }
        }
      }
    },
...
    "/plans": {
      "get": {
        "summary": "List all plans accessible to the user",
        "tags": [
          "Plans"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string"
            },
            "description": "Comma-separated plan statuses to filter by (e.g. \"active,draft\"). If omitted, returns all plans."
          }
        ],
        "responses": {
          "200": {
            "description": "A list of plans",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Plan"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          }
        }
      },
      "post": {
        "summary": "Create a new plan",
        "tags": [
          "Plans"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "title"
                ],
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "draft",
                      "active",
                      "completed",
                      "archived"
                    ]
                  },
                  "metadata": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Plan created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Plan"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Authentication required"
          }
        }
      }
    },
...
    "/plans/{id}/collaborators": {
      "get": {
        "summary": "List collaborators on a plan",
        "tags": [
          "Plans"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The plan ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Collaborators list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "owner": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "email": {
                          "type": "string"
                        }
                      }
                    },
                    "collaborators": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "user": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string"
                              },
                              "name": {
                                "type": "string"
                              },
                              "email": {
                                "type": "string"
                              }
                            }
                          },
                          "role": {
                            "type": "string"
                          },
                          "created_at": {
                            "type": "string",
                            "format": "date-time"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Access denied"
          },
          "404": {
            "description": "Plan not found"
          }
        }
      },
      "post": {
        "summary": "Add a collaborator to a plan",
        "tags": [
          "Plans"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The plan ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "role"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "role": {
                    "type": "string",
                    "enum": [
                      "viewer",
                      "editor",
                      "admin"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Collaborator role updated successfully"
          },
          "201": {
            "description": "Collaborator added successfully"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Access denied"
          },
          "404": {
            "description": "User or plan not found"
          }
        }
      }
    },
...
    "/plans/{id}/fork": {
      "post": {
        "summary": "Fork a plan into the caller's workspace",
        "description": "Deep-copies the plan, all plan_nodes (with translated parent_id),\nand dependency edges into a new plan owned by the caller.\nSource must be public OR the caller must already have access.\nLineage is recorded on metadata.forked_from / .forked_at.\n",
        "tags": [
          "Plans"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "New forked plan"
          },
          "403": {
            "description": "Caller lacks access to source"
          },
          "404": {
            "description": "Source plan not found"
          }
        }
      }
    },
...
    "/v1/plans": {
      "get": {
        "summary": "List plans accessible to the authenticated user",
        "tags": [
          "v1"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "Plan list"
          }
        }
      },
      "post": {
        "summary": "Create a plan",
        "tags": [
          "v1"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKey": []
          }
        ],
        "responses": {
          "201": {
            "description": "Plan created"
          }
        }
      }
    },
...
  },
...
}